+++ Source/Alternet.UI/Controls/BaseControls/SpeedButton/SpeedButton.Browsable.cs
@@ -1,0 +1,11 @@
+namespace Alternet.UI
+{
+    public partial class SpeedButton
+    {
+        internal new LayoutStyle? Layout
+        {
+            get => base.Layout;
+            set => base.Layout = value;
+        }
+    }
+}