+++ Source/Alternet.UI/Controls/TabControl/CardPanelHeader.Browsable.cs
@@ -1,0 +1,6 @@
+namespace Alternet.UI
+{
+    internal partial class CardPanelHeader
+    {
+    }
+}