+++ Source/Alternet.UI/Controls/ToolBar/GenericToolBar.Browsable.cs
@@ -1,0 +1,19 @@
+namespace Alternet.UI
+{
+    public partial class GenericToolBar
+    {
+        [Browsable(false)]
+        internal new string Text
+        {
+            get => base.Text;
+            set => base.Text = value;
+        }
+
+        [Browsable(false)]
+        internal new LayoutStyle? Layout
+        {
+            get => base.Layout;
+            set => base.Layout = value;
+        }
+    }
+}